requirements.txt
asgiref==3.3.1
boto3==1.17.2
botocore==1.20.2
Django==3.1.6
django-storages==1.11.1
jmespath==0.10.0
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
djangorestframework~=3.12.4
djangorestframework-jwt~=1.11.0
